+++ src/components/navigation/navigation.jsx
@@ -72,7 +72,7 @@ const Navigation = () => {
             flexDirection: "column",
             gap: 10,
             width: "100%",
-            pointerEvents: "auto"
+            pointerEvents: "auto",
           }}
         >
           <Button