+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Roboto Flex font added for ukrainian language (PATCH)
 - Cookie banner issue fixed (MINOR)
 - Fix dashboards' titles (PATCH)
+- Invoice/Product not found and page not found UI implemented (PATCH)
 
 # 1.3.0